Adds the JsonParser class.
The JsonParser is an entirely static class that takes a JSON string and 
returns an Action object that is a key-value store of information. JsonParser does not support hierarchical objects yet. It also requires that values that are represented by integers be casted to integers, as it internally stores numerical values as floats.